fix: partner create

+++ database/migrations/2025_11_19_064851_add_parent_id_to_categories_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('categories', function (Blueprint $table) {
-            $table->foreignId('parent_id')->nullable()->constrained('categories')->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('categories', 'parent_id')) {
+            Schema::table('categories', function (Blueprint $table) {
+                $table->foreignId('parent_id')->nullable()->constrained('categories')->onDelete('cascade');
+            });
+        }
     }
 
     /**
